+++ src/deion/worker/core/season/getDaysLeftSchedule.ts
@@ -6,7 +6,7 @@ import getSchedule from "./getSchedule";
  * @memberOf core.season
  * @return {Promise} The number of days left in the schedule.
  */
-const getDaysLeftSchedule = async (untilAllStarGame: boolean = false) => {
+const getDaysLeftSchedule = async (target?: number | "allStarGame") => {
 	const schedule = await getSchedule();
 	let numDays = 0;
 	let iPrev = 0;
@@ -17,18 +17,26 @@ const getDaysLeftSchedule = async (untilAllStarGame: boolean = false) => {
 		let i;
 
 		for (i = iPrev; i < schedule.length; i++) {
-			const { awayTid, homeTid } = schedule[i]; // All-Star Game
+			const { awayTid, homeTid, gid } = schedule[i]; // All-Star Game
 
 			if (awayTid === -2 && homeTid === -1) {
 				// Add one to represent that the "current day" of games before the ASG ended. Then the normal +1 at the end of the while loop will account for the ASG itself
 				numDays += 1;
 
-				if (untilAllStarGame) {
+				if (target === "allStarGame") {
 					return numDays;
 				}
 
 				i += 1;
 				break;
+			}
+
+			if (target === gid) {
+				// How many days are there before the day including this game? Need to add 1 if this game would be the first entry in a new day
+				if (tids.includes(homeTid) || tids.includes(awayTid)) {
+					numDays += 1;
+				}
+				return numDays;
 			}
 
 			if (!tids.includes(homeTid) && !tids.includes(awayTid)) {
@@ -41,6 +49,10 @@ const getDaysLeftSchedule = async (untilAllStarGame: boolean = false) => {
 
 		iPrev = i;
 		numDays += 1;
+	}
+
+	if (target !== undefined) {
+		throw new Error(`getDaysLeftSchedule did not find target "${target}"`);
 	}
 
 	return numDays;